requirements.txt
-i https://pypi.org/simple
django==2.1.7
djangorestframework==3.9.2
pytz==2018.9
